refactor: Add database migration functionality to main.clj
This commit adds the necessary code to perform database migrations in the `main.clj` file. It imports the `db-migration` namespace and calls the `db-up!` function to run the migrations. The commit also includes print statements to indicate the start and completion of the migration process.

Note: This commit message follows the convention of starting with a type (refactor) followed by a brief description of the changes.