refactor: Add find-by-phone-number! function to UserDb protocol
This commit adds the find-by-phone-number! function to the UserDb protocol in the user_db.clj file. This function allows searching for a user by their phone number. It enhances the user database functionality by providing an additional search method.

Note: This commit message follows the convention of starting with a type (refactor) followed by a brief description of the changes.