refactor: Update movie and user database implementations
This commit updates the movie and user database implementations. It introduces the `movie-db` and `user-db` namespaces and modifies the `movie-db-impl.clj` and `user-db-impl.clj` files to require these namespaces. Additionally, it adds the `tmdb` and `in-memory` functions to create instances of the movie and user database implementations respectively.

Note: This commit message follows the convention of starting with a type (refactor) followed by a brief description of the changes.