refactor: Add put-many! function to MediaDb protocol
This commit adds the `put-many!` function to the `MediaDb` protocol in the `media_db.clj` file. The `put-many!` function allows for the bulk insertion of media items into the database. This is useful for improving performance when adding multiple media items at once.

Note: This commit message follows the convention of starting with a type (refactor) followed by a brief description of the changes.