refactor: Remove movie and media database implementations
This commit removes the movie and media database implementations, including the `movie-db` and `media-db` namespaces, as well as the `movie-db-impl.clj` and `media-db-impl.clj` files. These changes are no longer needed and simplify the codebase.

Note: This commit message follows the convention of starting with a type (refactor) followed by a brief description of the changes.