refactor: Rename tmdb->movie functions to tmdb-movie->media
This commit renames the `tmdb->movie` functions to `tmdb-movie->media` to better reflect their purpose of converting TMDB movie data to media data. The function names are updated in the `media_db_impl_tmdb_movie.clj` file. This change improves code clarity and consistency.

Note: This commit message follows the convention of starting with a type (refactor) followed by a brief description of the changes.