refactor: Improve image lazy loading and remove unused code
This commit refactors the moviefinder_app/home.clj file to improve the image lazy loading functionality. It removes the unused view-youtube-video function, which was commented out. These changes optimize the performance of the application and remove unnecessary code.

Note: This commit message follows the convention of starting with a type (refactor) followed by a brief description of the changes.