refactor: Improve image lazy loading and update Twilio API usage
This commit refactors the code in the moviefinder_app.login.login-with-sms.verify-sms.verify-sms-impl-twilio namespace to improve the image lazy loading functionality. It also updates the usage of the Twilio API to use the http-client namespace instead of clj-http.client. These changes optimize the performance of the application and ensure compatibility with the latest dependencies.

Note: This commit message follows the convention of starting with a type (refactor) followed by a brief description of the changes.